# Conflicts:
#	inventory/src/main/java/org/opennms/horizon/inventory/service/MonitoringLocationService.java

+++ minion/plugins/plugins-api/src/main/java/org/opennms/horizon/minion/plugin/api/MonitoredService.java
@@ -57,7 +57,7 @@ public interface MonitoredService {
      *
      * @return the nodeid
      */
-    int getNodeId();
+    long getNodeId();
 
     /**
      * Returns the label of the node that this service is associated with.
